+++ Plugins/HelpDesk.Plugin/Models/FaqGroupItem.cs
@@ -33,7 +33,7 @@ namespace HelpdeskPlugin.Models
     {
         #region Constructors
 
-        public FaqGroupItem(in int id, in string description,
+        public FaqGroupItem(in long id, in string description,
             in int viewCount, in string content)
         {
             if (String.IsNullOrEmpty(description))
@@ -55,7 +55,7 @@ namespace HelpdeskPlugin.Models
 
         #region Properties
 
-        public int Id { get; private set; }
+        public long Id { get; private set; }
 
         public string Description { get; private set; }
